Split the cxx indexer lib target into several smaller targets to make dependencies clearer and improve recompilation times.
