+++ kythe/cxx/indexer/cxx/IndexerLibrarySupport.h
@@ -97,28 +97,6 @@ class LibrarySupport {
 /// \brief A collection of library support implementations.
 using LibrarySupports = std::vector<std::unique_ptr<LibrarySupport>>;
 
-/// \brief Snoops on variable declarations and references to see if they
-/// are flags.
-class GoogleFlagsLibrarySupport : public LibrarySupport {
- public:
-  GoogleFlagsLibrarySupport() {}
-
-  /// \brief Emits a google/gflag node if `Decl` is a flag.
-  void InspectVariable(IndexerASTVisitor &V, GraphObserver::NodeId &DeclNodeId,
-                       GraphObserver::NodeId &DeclBodyNodeId,
-                       const clang::VarDecl *Decl,
-                       GraphObserver::Completeness Compl,
-                       const std::vector<Completion> &Compls) override;
-
-  /// \brief Checks whether this DeclRef refers to a flag. If so, it emits an
-  /// additional ref edge to the correspondng google/gflag node.
-  void InspectDeclRef(IndexerASTVisitor &V,
-                      clang::SourceLocation DeclRefLocation,
-                      const GraphObserver::Range &Ref,
-                      GraphObserver::NodeId &RefId,
-                      const clang::NamedDecl *TargetDecl) override;
-};
-
 }  // namespace kythe
 
 #endif  // KYTHE_CXX_INDEXER_CXX_INDEXER_LIBRARY_SUPPORT_H_
